LLM Remotely translatable specs changes
Contains the adjustments we discussed for the tests related to spec/shared/system/remotely_translatable.rb, along with a few simplifications and some alignment with our component conventions.
For the remotely translatable changes, I opted to allow a new provider and make one call for :microsoft and another for :llm. Since the Microsoft option will likely be deprecated in upcoming releases, we can revisit and optimize those tests later on.